moim_detaill_ejs

+++ controller/Cmoim.js
@@ -9,16 +9,19 @@ const {
 
 exports.MoimList_GET = async (req, res) => {
   res.render("moim_list");
-}
+};
 
 exports.Moims_GET = async (req, res) => {
   try {
     const data = await Moim.findAll();
     console.log(data);
-    
+
     res.json({ data: data });
-  } catch(error){
-    res.json({ result: false, Message: "모임 정보 불러오기에 실패하였습니다!!!" });
+  } catch (error) {
+    res.json({
+      result: false,
+      Message: "모임 정보 불러오기에 실패하였습니다!!!",
+    });
   }
 };
 
@@ -63,40 +66,40 @@ exports.Moimset_patch = async (req, res) => {
 };
 
 exports.MoimSet_detory = async (req, res) => {
-  // if (req.session.userInfo) {
-  try {
-    const { user_review, moim_id, user_id } = req.body;
-    await MoimSet.destroy({ where: { moim_id, user_id } });
-    res.json({ result: true, Message: "모임 가입을 취소하였습니다." });
-  } catch (error) {
-    res.send({
-      result: false,
-      Message: "에러 발생!! 모임 가입을 해제할 수 없습니다.",
-    });
+  if (req.session.userInfo) {
+    try {
+      const { user_review, moim_id, user_id } = req.body;
+      await MoimSet.destroy({ where: { moim_id, user_id } });
+      res.json({ result: true, Message: "모임 가입을 취소하였습니다." });
+    } catch (error) {
+      res.send({
+        result: false,
+        Message: "에러 발생!! 모임 가입을 해제할 수 없습니다.",
+      });
+    }
+  } else {
+    res.redirect("/login");
   }
-  // } else {
-  //   res.redirect("/login");
-  // }
 };
 
 exports.MoimSet_POST = async (req, res) => {
-  // if (req.session.userInfo) {
-  try {
-    const { user_review, moim_id, user_id } = req.body;
-    MoimSet.create({ moim_id, user_id });
-    res.send({
-      result: true,
-      Message: "모임에 가입해주신 것을 환영합니다.",
-    });
-  } catch (error) {
-    res.send({
-      result: false,
-      Message: "에러 발생!! 모임에 가입할 수 없습니다.",
-    });
+  if (req.session.userInfo) {
+    try {
+      const { moim_id } = req.body;
+      MoimSet.create({ moim_id, user_id: req.session.userInfo.userid });
+      res.json({
+        result: true,
+        Message: "모임에 가입해주신 것을 환영합니다.",
+      });
+    } catch (error) {
+      res.json({
+        result: false,
+        Message: `${error} 에러 발생!! 모임에 가입할 수 없습니다.`,
+      });
+    }
+  } else {
+    req.redirect("/login");
   }
-  // } else {
-  //   req.redirect("/login");
-  // }
 };
 
 exports.moim_detail_UPDATE = async (req, res) => {
@@ -111,20 +114,8 @@ exports.moim_detail_UPDATE = async (req, res) => {
 };
 
 exports.Moim_UPDATE = async (req, res) => {
-  // if (req.session.userInfo) {
-  const {
-    title,
-    on_line,
-    max_people,
-    expiration_date,
-    even_date,
-    location,
-    represent_img,
-    user_id,
-    moim_id,
-  } = req.body;
-  await Moim.update(
-    {
+  if (req.session.userInfo) {
+    const {
       title,
       on_line,
       max_people,
@@ -133,131 +124,145 @@ exports.Moim_UPDATE = async (req, res) => {
       location,
       represent_img,
       user_id,
-    },
-    { where: { moim_id } }
-  );
-  res.send({
-    result: true,
-    Message: "moim 정보 업데이트 1단계에 성공하셨습니다.",
-  });
-  // } else {
-  //   res.redirect("/login");
-  // }
+      moim_id,
+    } = req.body;
+    await Moim.update(
+      {
+        title,
+        on_line,
+        max_people,
+        expiration_date,
+        even_date,
+        location,
+        represent_img,
+        user_id,
+      },
+      { where: { moim_id } }
+    );
+    res.send({
+      result: true,
+      Message: "moim 정보 업데이트 1단계에 성공하셨습니다.",
+    });
+  } else {
+    res.redirect("/login");
+  }
 };
 
 exports.MoimDetail_POST = async (req, res) => {
-  // if (req.session.userInfo) {
-  const { moim_id, content, min_people } = req.body;
-  try {
-    await MoimDetail.create({ moim_id, content, min_people });
-    res.json({ result: true });
-  } catch (error) {
-    console.error(error);
-    await Moim.destroy({ where: { moim_id } });
-    //Moim_detaill 테이블에 정보 저장이 실패하였을 때, Moim table의 이전 저장 정보를 삭제한다.
-    res.send({ result: false, Message: "모임 개설에 실패하였습니다." });
+  if (req.session.userInfo) {
+    const { moim_id, content, min_people } = req.body;
+    try {
+      await MoimDetail.create({ moim_id, content, min_people });
+      res.json({ result: true });
+    } catch (error) {
+      console.error(error);
+      await Moim.destroy({ where: { moim_id } });
+      //Moim_detaill 테이블에 정보 저장이 실패하였을 때, Moim table의 이전 저장 정보를 삭제한다.
+      res.send({ result: false, Message: "모임 개설에 실패하였습니다." });
+    }
+  } else {
+    res.redirect("/login");
   }
-  // } else {
-  //   res.redirect("/login");
-  // }
 };
 
 exports.Moims_POST = async (req, res) => {
-  // if (req.session.userInfo) {
-  try {
-    const {
-      category,
-      title,
-      on_line,
-      max_people,
-      expiration_date,
-      even_date,
-      location,
-      represent_img,
-      user_id,
-    } = req.body;
+  if (req.session.userInfo) {
+    try {
+      const {
+        category,
+        title,
+        on_line,
+        max_people,
+        expiration_date,
+        even_date,
+        location,
+        represent_img,
+        user_id,
+      } = req.body;
 
-    console.log(req.body);
-    const date = await Moim.create({
-      title,
-      category,
-      on_line,
-      max_people,
-      expiration_date,
-      even_date,
-      location,
-      represent_img,
-      user_id,
-    });
-    res.json({ result: true, userInfo: date });
-  } catch (error) {
-    console.error(error);
-    res.send({
-      result: false,
-      Message: "모임 개설에 실패하였습니다.",
-      userInfo: null,
-    });
+      console.log(req.body);
+      const date = await Moim.create({
+        title,
+        category,
+        on_line,
+        max_people,
+        expiration_date,
+        even_date,
+        location,
+        represent_img,
+        user_id,
+      });
+      res.json({ result: true, userInfo: date });
+    } catch (error) {
+      console.error(error);
+      res.send({
+        result: false,
+        Message: "모임 개설에 실패하였습니다.",
+        userInfo: null,
+      });
+    }
+  } else {
+    res.redirect("/login");
   }
-  // } else {
-  //   res.redirect("/login");
-  // }
 };
 
 // 모임 디테일 페이지 렌더링
 exports.MoimDetail_render = async (req, res) => {
   console.log(req.params.moimid);
-  try {
-    const data = await Moim.findOne({where: {moim_id: req.params.moimid}});
-    // const detail = await MoimDetail.findOne({where: {moim_id: req.params.moim_id}})
-    res.render("moim_detail", {data});
-  } catch(error){
-    res.json({ result: true, Message: "모임 정보 불러오기에 실패하였습니다!!!" });
-  }
+  res.render("moim_detail");
 };
 
 exports.moimlist = async (req, res) => {
-  const data = await Moim.findAll({
-    attributes: [
-      "title",
-      "on_line",
-      "max_people",
-      "location",
-      "represent_img",
-      "user_id",
-      "category",
-      [
-        sequelize.fn(
-          "date_format",
-          sequelize.col("expiration_date"),
-          "%Y-%d-%m %H:%i"
-        ),
-        "expiration_date",
-      ],
-      [
-        sequelize.fn(
-          "date_format",
-          sequelize.col("even_date"),
-          "%Y-%d-%m %H:%i"
-        ),
-        "even_date",
-      ],
-    ],
-  });
-  const moimset = await MoimSet.findAll({
-    attributes: [
-      "moim_id",
-      [sequelize.fn("COUNT", sequelize.col("user_id")), "moim_count"],
-    ],
-    group: "moim_id",
-  });
-  let moimcount = [];
-  for (let i = 0; i < moimset.length; i++) {
-    moimcount.push(moimset[i].dataValues);
-  }
-  if (data) {
-    res.render("moimlist", { data, moimcount });
+  if (req.session.userInfo) {
+    try {
+      const data = await Moim.findAll({
+        attributes: [
+          "title",
+          "on_line",
+          "max_people",
+          "location",
+          "represent_img",
+          "user_id",
+          "category",
+          [
+            sequelize.fn(
+              "date_format",
+              sequelize.col("expiration_date"),
+              "%Y-%d-%m %H:%i"
+            ),
+            "expiration_date",
+          ],
+          [
+            sequelize.fn(
+              "date_format",
+              sequelize.col("even_date"),
+              "%Y-%d-%m %H:%i"
+            ),
+            "even_date",
+          ],
+        ],
+      });
+      const moimset = await MoimSet.findAll({
+        attributes: [
+          "moim_id",
+          [sequelize.fn("COUNT", sequelize.col("user_id")), "moim_count"],
+        ],
+        group: "moim_id",
+      });
+      let moimcount = [];
+      for (let i = 0; i < moimset.length; i++) {
+        moimcount.push(moimset[i].dataValues);
+      }
+      if (data) {
+        res.render("moimlist", { data, moimcount });
+      } else {
+        alert("모임 리스트 출력 실패");
+        res.redirect("/");
+      }
+    } catch (error) {
+      console.error(error);
+    }
   } else {
-    alert("모임 리스트 출력 실패");
-    res.redirect("/");
+    res.redirect("/login");
   }
 };
